Find and remove the fragmented Huffman chunks
Previously the Huffman modules were removed by wiping all the data between
the first and the last chunk. Unfortunately the modules are often
fragmented: the chunks are not in a linear order, sometimes repeated and
sometimes shared between different modules.

This commit removes each chunk individually, after checking it's not used
by a fundamental module.