Do not set the HAP/AltMeDisable bit by default
While some BIOS behaves better with the bit set, others don't. As
me_cleaner has been already tested on many platforms (and we don't want to
break things if it previously worked), revert to the default behaviour
and add an option (-S) to optionally set the HAP/AltMeDisable bit.

So now the situation is:
 -s: just set the HAP/AltMeDisable bit, without modifying the firmware
 -S: modify the firmware and set the HAP/AltMeDisable bit
 none of them: modify the firmware but don't set the HAP/AltMeDisable bit